Faster NA-replacement in aspcf
The code to replace NAs with adjacent values in ascat.aspcf ended up being one of the slower parts of the entire function for me - presumably my data has a lot of NAs. This separates the NA-infill to a new, faster, function. Note also that the code it replaces seems to be subtly wrong (closestnonna is a position in nonnakes, not in logRPCFed ).